1_Q&Achatbot/requirements.txt
python-dotenv
langchain
langchain_community
langchain_core
streamlit
openai
langchain_openai
huggingface_hub
langchain_groq
langchain_ollama